src/scripts/requirements.txt
mysql-connector-python==9.3.0
pandas==2.2.3
scikit-learn==1.6.1
numpy==2.2.5
nltk==3.9.1
